requirements.txt
click==8.0.1
flask==2.0.1
flask-Cors==3.0.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
six==1.16.0
# uWSGI==2.0.19.1
Werkzeug==2.0.1
python-dotenv==0.19.0
boto3